Removed JSON id's
The id property of the JSON elements is unnecessary, because you can use the index property when iterating over the JSON in React. This makes it easier to rearrange things as well, because they will be displayed in whatever order you have them in the JSON, rather than you having to change all the id's to experiment.